server/requirements.txt
Django==4.2.15
whitenoise==6.6.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.3.0
django-jazzmin==3.0.0
django-tailwind==3.8.0
Django-Verify-Email==2.0.3